split acceptance tests builds to not hit travis job timeout
